EXPERIMENTAL: Cleanup + Remove PPLib
I just cleaned up some imports and removed PPLib for good. The only problem is that for CustomNamedCommands, I used the wrappedEventCommand to make each command unique and thus not cause issues of registering the same command twice in a sequential group. I copied his EXACT code and put it in RobotContainer UNTESTED